big changed of folder structure to classic `src` - makes tracing of changes hard :( - added `common` with base_model *for pydantic) - and cognite_client   - should be the same for  all projects   - sadly bootstrap/extpipes/transformations use an idp-subsection, instead of a flat properties `cognite` config - sharing same code structure with next extpipes-cli v3
