+++ pyproject.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 repository = "https://github.com/cognitedata/inso-bootstrap-cli"
 
 packages = [
-    { include="incubator", from="." },
+    { include="src" },
 ]
 [tool.poetry.group.dev.dependencies]
 autopep8 = "^2.0.1"
@@ -17,7 +17,7 @@ autopep8 = "^2.0.1"
 
 [tool.black]
 line-length = 120
-target_version = ['py39']
+target_version = ['py310']
 include = '\.py$'
 
 [tool.isort]
@@ -53,7 +53,7 @@ pytest = "^7.2.1"
 # pytest-cov = "^3.0.0"
 
 [tool.poetry.scripts]
-bootstrap-cli = "incubator.bootstrap_cli.__main__:main"
+bootstrap-cli = "src.__main__:main"
 
 [build-system]
 requires = ["poetry>=0.12"]
@@ -64,5 +64,5 @@ version_toml = [
 "pyproject.toml:tool.poetry.version"
 ]
 version_variable = [
-    "incubator/bootstrap_cli/__init__.py:__version__",
+    "src/__init__.py:__version__",
 ]
