+++ src/commands/delete.py
