+++ src/commands/diagram.py
